cpp + regex =  wuuuuuuuuuuuuuuuuuuut

+++ includes/IOperand.hpp
@@ -19,7 +19,7 @@ private:
 protected:
 
 public:
-	static int const			precisions[5];
+	static double const			precisions[5];
 
 	virtual int						getPrecision( void ) const = 0;
 	virtual eOperandType	getType( void ) const = 0;
